Refactor to use React Hooks and Context API
