+++ client/components/Search.jsx
@@ -1,14 +1,17 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import styled from 'styled-components';
+import { StateContext } from './contexts/StateContext.jsx';
 
-export default function Search({ handleSearch }) {
+export default function Search() {
+  const { handleSearchCard } = useContext(StateContext);
+
   return (
     <div>
-      <Form onSubmit={handleSearch}>
+      <Form onSubmit={handleSearchCard}>
         <label htmlFor="card-search">Search Card Name: </label>
         <input type="search" id="card-search" placeholder="Card Name" />
       </Form>
-      <Go onClick={handleSearch}>Go</Go>
+      <Go onClick={handleSearchCard}>Go</Go>
     </div>
   )
 }
